requirements.txt
# Web framework
flask>=3.0.0

# MCP
mcp>=1.0.0

# GitHub + HTTP
requests>=2.31.0

# CodeBERT embeddings
transformers>=4.38.0
torch>=2.1.0

# FAISS
faiss-cpu>=1.7.4

# Tree-sitter AST parsing
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0

# Utilities
numpy>=1.26.0
python-dotenv>=1.0.0
